[change] rename modules & enums
