+++ ata_api/helpers/functools.py
