+++ ata_api/models.py
@@ -3,7 +3,7 @@ from uuid import UUID
 from ata_db_models.models import Group
 from pydantic import BaseModel
 
-from ata_api.helpers.enums import SiteName
+from ata_api.site import SiteName
 
 
 class PrescriptionResponse(BaseModel):
