+++ ata_api/site.py
@@ -1,0 +1,10 @@
+from enum import auto
+
+from ata_api.helpers.enum import StrEnumKebab
+
+
+class SiteName(StrEnumKebab):
+    THE_19TH = auto()
+    OPEN_VALLEJO = auto()
+    DALLAS_FREE_PRESS = auto()
+    AFRO_LA = auto()
