+++ pyproject.toml
@@ -13,6 +13,8 @@ ata-db-models = "0.0.17"
 fastapi = "^0.95.0"
 mangum = "^0.17.0"
 pydantic = "^1.10.8"
+aws-lambda-powertools = "^2.16.1"
+case-converter = "^1.1.0"
 
 [tool.poetry.group.dev.dependencies]
 mypy = "^0.961"
